// extensions/first-order-limit/generated/input.ts
export type CartValidationsGenerateRunInput = { cart: { lines: Array<{ id: string }>, buyerIdentity: { customer: { numberOfOrders: number } | null } | null, cost: { subtotalAmount: { amount: string, currencyCode: CurrencyCode } } }, validation: { config: { jsonValue: unknown } | null } };